JSSpecCompiler: Move passes to Passes subdirectory of Compiler/
